refactor: consolidate event system API for consistency (#989)
Standardize event-related types and utilities by consolidating duplicated
code into a single unified implementation in events.ts.

Changes:
- Move GetEntityEventBus type from eventBubbling/eventDescendants to events.ts
- Move EntityEventBusStore interface to events.ts
- Move createEntityEventBusStore function to events.ts with unified signature
- Update all imports across codebase to reference events.ts
- Update tests to use unified API signatures
- Update public exports in index.ts and types/index.ts

This ensures consistent naming, parameter order, and return types across
the event system API.
