+++ src/core/events.ts
@@ -3,7 +3,7 @@
  * @module core/events
  */
 
-import type { Unsubscribe } from './types';
+import type { Entity, Unsubscribe, World } from './types';
 
 /**
  * Generic event handler type.
@@ -194,4 +194,114 @@ export function createEventBus<T extends EventMap>(): EventBus<T> {
 	};
 
 	return bus;
+}
+
+// =============================================================================
+// ENTITY EVENT BUS UTILITIES
+// =============================================================================
+
+/**
+ * Function type for getting an EventBus for a specific entity.
+ * Returns undefined if the entity has no event bus.
+ *
+ * @typeParam T - Event map type
+ * @param world - The ECS world
+ * @param eid - The entity ID
+ * @returns The entity's EventBus, or undefined if none exists
+ *
+ * @example
+ * ```typescript
+ * import { createEventBus, type GetEntityEventBus } from 'blecsd';
+ *
+ * const entityBuses = new Map<Entity, EventBus<MyEvents>>();
+ * const getEventBus: GetEntityEventBus<MyEvents> = (world, eid) => entityBuses.get(eid);
+ * ```
+ */
+export type GetEntityEventBus<T extends EventMap> = (
+	world: World,
+	eid: Entity,
+) => EventBus<T> | undefined;
+
+/**
+ * Store for entity event buses.
+ * Maps entity IDs to their event buses.
+ *
+ * @typeParam T - Event map type
+ */
+export interface EntityEventBusStore<T extends EventMap> {
+	/** Gets the event bus for an entity */
+	get(world: World, eid: Entity): EventBus<T> | undefined;
+	/** Gets or creates an event bus for an entity */
+	getOrCreate(world: World, eid: Entity): EventBus<T>;
+	/** Sets the event bus for an entity */
+	set(world: World, eid: Entity, eventBus: EventBus<T>): void;
+	/** Checks if an entity has an event bus */
+	has(world: World, eid: Entity): boolean;
+	/** Removes the event bus for an entity */
+	delete(world: World, eid: Entity): boolean;
+	/** Clears all event buses */
+	clear(): void;
+}
+
+/**
+ * Creates a simple entity event bus store using a Map.
+ * Provides a centralized way to manage EventBus instances per entity.
+ *
+ * @typeParam T - Event map type
+ * @returns A new EntityEventBusStore instance
+ *
+ * @example
+ * ```typescript
+ * import { createEntityEventBusStore, createEventBus } from 'blecsd';
+ *
+ * interface MyEvents {
+ *   click: { x: number; y: number };
+ *   focus: Record<string, never>;
+ * }
+ *
+ * const store = createEntityEventBusStore<MyEvents>();
+ *
+ * // Get or create a bus for an entity
+ * const bus = store.getOrCreate(world, entityId);
+ * bus.on('click', (e) => console.log('clicked at', e.x, e.y));
+ *
+ * // Check if entity has a bus
+ * if (store.has(world, entityId)) {
+ *   const existingBus = store.get(world, entityId);
+ * }
+ * ```
+ */
+export function createEntityEventBusStore<T extends EventMap>(): EntityEventBusStore<T> {
+	const buses = new Map<Entity, EventBus<T>>();
+
+	return {
+		get(_world: World, eid: Entity): EventBus<T> | undefined {
+			return buses.get(eid);
+		},
+
+		getOrCreate(_world: World, eid: Entity): EventBus<T> {
+			let bus = buses.get(eid);
+			if (!bus) {
+				bus = createEventBus<T>();
+				buses.set(eid, bus);
+			}
+			return bus;
+		},
+
+		set(_world: World, eid: Entity, eventBus: EventBus<T>): void {
+			buses.set(eid, eventBus);
+		},
+
+		has(_world: World, eid: Entity): boolean {
+			return buses.has(eid);
+		},
+
+		delete(_world: World, eid: Entity): boolean {
+			return buses.delete(eid);
+		},
+
+		clear(): void {
+			buses.clear();
+		},
+	};
 }
